Bug 572897 lic api goes out from the shadows
publish
 - `lic.api.access`
 - `Access Cycle` implementation
