+++ bundles/org.eclipse.passage.lac/src/org/eclipse/passage/internal/lac/base/PassageAgent.java
@@ -12,10 +12,10 @@
  *******************************************************************************/
 package org.eclipse.passage.internal.lac.base;
 
+import org.eclipse.passage.lic.api.access.GrantLockAttempt;
 import org.eclipse.passage.lic.internal.api.LicensedProduct;
 import org.eclipse.passage.lic.internal.api.Passage;
 import org.eclipse.passage.lic.internal.api.ServiceInvocationResult;
-import org.eclipse.passage.lic.internal.api.access.GrantLockAttempt;
 import org.eclipse.passage.lic.internal.api.restrictions.ExaminationCertificate;
 import org.eclipse.passage.lic.internal.equinox.EquinoxPassage;
 import org.eclipse.passage.lic.internal.net.NetFrameworkAware;
