+++ bundles/org.eclipse.passage.lic.api/src/org/eclipse/passage/lic/api/io/KeyKeeperRegistry.java
@@ -17,6 +17,9 @@ import java.util.function.Supplier;
 import org.eclipse.passage.lic.api.registry.Registry;
 import org.eclipse.passage.lic.internal.api.LicensedProduct;
 
+/**
+ * @since 1.1
+ */
 public interface KeyKeeperRegistry extends Supplier<Registry<LicensedProduct, KeyKeeper>> {
 
 }
