+++ bundles/org.eclipse.passage.lic.api/src/org/eclipse/passage/lic/api/registry/Service.java
@@ -20,6 +20,7 @@ package org.eclipse.passage.lic.api.registry;
  * </p>
  * 
  * @param <I> sub type of {@linkplain ServiceId}
+ * @since 1.1
  */
 public interface Service<I extends ServiceId> {
