+++ bundles/org.eclipse.passage.lic.api/src/org/eclipse/passage/lic/api/registry/StringServiceId.java
@@ -17,6 +17,7 @@ import java.util.Objects;
 /**
  * Base {@linkplain ServiceId} service implementation represents a single
  * {@code not null} string.
+ * @since 1.1
  */
 public final class StringServiceId implements ServiceId {
