+++ bundles/org.eclipse.passage.lic.api/src/org/eclipse/passage/lic/internal/api/PassageUI.java
@@ -12,7 +12,7 @@
  *******************************************************************************/
 package org.eclipse.passage.lic.internal.api;
 
-import org.eclipse.passage.lic.internal.api.access.GrantLockAttempt;
+import org.eclipse.passage.lic.api.access.GrantLockAttempt;
 import org.eclipse.passage.lic.internal.api.restrictions.ExaminationCertificate;
 
 /**
