+++ bundles/org.eclipse.passage.lic.base/META-INF/MANIFEST.MF
@@ -2,14 +2,15 @@ Manifest-Version: 1.0
 Automatic-Module-Name: org.eclipse.passage.lic.api.base
 Bundle-ManifestVersion: 2
 Bundle-SymbolicName: org.eclipse.passage.lic.base
-Bundle-Version: 1.0.300.qualifier
+Bundle-Version: 1.1.0.qualifier
 Bundle-Name: %Bundle-Name
 Bundle-Vendor: %Bundle-Vendor
 Bundle-Copyright: %Bundle-Copyright
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Require-Bundle: org.eclipse.passage.lic.api;bundle-version="1.0.0";visibility:=reexport,
  org.apache.logging.log4j;bundle-version="2.8.2"
-Export-Package: org.eclipse.passage.lic.base.io,
+Export-Package: org.eclipse.passage.lic.base.access,
+ org.eclipse.passage.lic.base.io,
  org.eclipse.passage.lic.base.registry,
  org.eclipse.passage.lic.internal.base;
   x-friends:="org.eclipse.passage.lac,
@@ -27,7 +28,6 @@ Export-Package: org.eclipse.passage.lic.base.io,
    org.eclipse.passage.loc.products.emfforms,
    org.eclipse.passage.seal.demo,
    org.eclipse.passage.seal.demo.tests",
- org.eclipse.passage.lic.internal.base.access;x-friends:="org.eclipse.passage.seal.demo.tests,org.eclipse.passage.lic.equinox",
  org.eclipse.passage.lic.internal.base.acquire;x-friends:="org.eclipse.passage.lic.floating.model,org.eclipse.passage.lic.equinox,org.eclipse.passage.lic.licenses.model",
  org.eclipse.passage.lic.internal.base.conditions;
   x-friends:="org.eclipse.passage.lbc.base,
