+++ bundles/org.eclipse.passage.lic.base/src/org/eclipse/passage/lic/base/access/Allow.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -10,33 +10,34 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.internal.base.access;
+package org.eclipse.passage.lic.base.access;
 
-import java.util.function.Predicate;
-
-import org.eclipse.passage.lic.internal.api.ServiceInvocationResult;
+import org.eclipse.passage.lic.internal.api.Framework;
+import org.eclipse.passage.lic.internal.api.diagnostic.Diagnostic;
 import org.eclipse.passage.lic.internal.api.restrictions.ExaminationCertificate;
 import org.eclipse.passage.lic.internal.base.diagnostic.NoSevereErrors;
 import org.eclipse.passage.lic.internal.base.restrictions.NoSevereRestrictions;
 
-/**
- * Tells if the examination result should cause use notification
- */
-public final class ShouldBeExposed implements Predicate<ServiceInvocationResult<ExaminationCertificate>> {
+final class Allow extends Cycle<Boolean> {
+
+	Allow(Framework framework, String feature) {
+		super(framework, feature);
+	}
 
 	@Override
-	public boolean test(ServiceInvocationResult<ExaminationCertificate> result) {
-		if (!result.data().isPresent()) {
-			return true;
-		}
-		if (!new NoSevereErrors().test(result.diagnostic())) {
-			return true;
-		}
-
-		if (!new NoSevereRestrictions().test(result.data().get())) {
-			return true;
-		}
+	protected Boolean stopOnError(Diagnostic diagnostic) {
 		return false;
+	}
+
+	@Override
+	protected Boolean stopOnCertificate(ExaminationCertificate certificate, Diagnostic diagnostic) {
+		return new NoSevereErrors().test(diagnostic) && //
+				new NoSevereRestrictions().test(certificate);
+	}
+
+	@Override
+	protected Boolean freeWayOut() {
+		return true;
 	}
 
 }
