+++ bundles/org.eclipse.passage.lic.base/src/org/eclipse/passage/lic/base/access/CanProceed.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -10,34 +10,25 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.internal.base.access;
+package org.eclipse.passage.lic.base.access;
 
-import org.eclipse.passage.lic.internal.api.Framework;
-import org.eclipse.passage.lic.internal.api.diagnostic.Diagnostic;
+import java.util.function.Predicate;
+
+import org.eclipse.passage.lic.internal.api.ServiceInvocationResult;
 import org.eclipse.passage.lic.internal.api.restrictions.ExaminationCertificate;
 import org.eclipse.passage.lic.internal.base.diagnostic.NoSevereErrors;
 import org.eclipse.passage.lic.internal.base.restrictions.NoSevereRestrictions;
 
-final class Allow extends Cycle<Boolean> {
-
-	Allow(Framework framework, String feature) {
-		super(framework, feature);
-	}
-
-	@Override
-	protected Boolean stopOnError(Diagnostic diagnostic) {
-		return false;
-	}
+/**
+ * @since 1.1
+ */
+public final class CanProceed implements Predicate<ServiceInvocationResult<ExaminationCertificate>> {
 
 	@Override
-	protected Boolean stopOnCertificate(ExaminationCertificate certificate, Diagnostic diagnostic) {
-		return new NoSevereErrors().test(diagnostic) && //
-				new NoSevereRestrictions().test(certificate);
-	}
-
-	@Override
-	protected Boolean freeWayOut() {
-		return true;
+	public boolean test(ServiceInvocationResult<ExaminationCertificate> result) {
+		return new NoSevereErrors().test(result.diagnostic()) && //
+				result.data().map(certificate -> new NoSevereRestrictions().test(certificate))//
+						.orElse(false);
 	}
 
 }
