+++ bundles/org.eclipse.passage.lic.base/src/org/eclipse/passage/lic/base/access/Permissions.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -10,7 +10,7 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.internal.base.access;
+package org.eclipse.passage.lic.base.access;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -32,6 +32,8 @@ import org.eclipse.passage.lic.internal.base.SumOfCollections;
 
 /**
  * FIXME: Has public visibility only for testing.
+ * 
+ * @since 1.1
  */
 public final class Permissions implements Supplier<ServiceInvocationResult<Collection<Permission>>> {
