+++ bundles/org.eclipse.passage.lic.base/src/org/eclipse/passage/lic/base/access/Requirements.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -10,7 +10,7 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.internal.base.access;
+package org.eclipse.passage.lic.base.access;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -34,6 +34,9 @@ import org.eclipse.passage.lic.internal.base.diagnostic.code.NoServicesOfType;
 import org.eclipse.passage.lic.internal.base.i18n.BaseMessages;
 import org.eclipse.passage.lic.internal.base.requirements.RequirementsFeatureFilter;
 
+/**
+ * @since 1.1
+ */
 public final class Requirements implements Supplier<ServiceInvocationResult<Collection<Requirement>>> {
 
 	private final Registry<StringServiceId, ResolvedRequirements> registry;
