+++ bundles/org.eclipse.passage.lic.base/src/org/eclipse/passage/lic/base/access/SumOfCertificates.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 ArSysOp
+ * Copyright (c) 2020, 2021 ArSysOp
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -10,7 +10,7 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.internal.base.access;
+package org.eclipse.passage.lic.base.access;
 
 import java.util.Map;
 import java.util.function.BinaryOperator;
@@ -25,6 +25,9 @@ import org.eclipse.passage.lic.internal.base.SumOfCollections;
 import org.eclipse.passage.lic.internal.base.SumOfMaps;
 import org.eclipse.passage.lic.internal.base.restrictions.BaseExaminationCertificate;
 
+/**
+ * @since 1.1
+ */
 public final class SumOfCertificates implements BinaryOperator<ExaminationCertificate> {
 
 	@Override
