+++ bundles/org.eclipse.passage.lic.base/src/org/eclipse/passage/lic/base/io/UserHomePath.java
@@ -16,6 +16,9 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.function.Supplier;
 
+/**
+ * @since 1.1
+ */
 public final class UserHomePath implements Supplier<Path> {
 
 	@Override
