+++ bundles/org.eclipse.passage.lic.equinox/src/org/eclipse/passage/lic/internal/equinox/EquinoxPassageLicenseCoverage.java
@@ -12,10 +12,10 @@
  *******************************************************************************/
 package org.eclipse.passage.lic.internal.equinox;
 
+import org.eclipse.passage.lic.base.access.Access;
 import org.eclipse.passage.lic.internal.api.PassageLicenseCoverage;
 import org.eclipse.passage.lic.internal.api.ServiceInvocationResult;
 import org.eclipse.passage.lic.internal.api.restrictions.ExaminationCertificate;
-import org.eclipse.passage.lic.internal.base.access.Access;
 
 public final class EquinoxPassageLicenseCoverage implements PassageLicenseCoverage {
