+++ bundles/org.eclipse.passage.lic.equinox/src/org/eclipse/passage/lic/internal/equinox/LicensedRunnable.java
@@ -15,8 +15,8 @@ package org.eclipse.passage.lic.internal.equinox;
 import java.util.Optional;
 import java.util.function.Consumer;
 
+import org.eclipse.passage.lic.api.access.GrantLockAttempt;
 import org.eclipse.passage.lic.internal.api.ServiceInvocationResult;
-import org.eclipse.passage.lic.internal.api.access.GrantLockAttempt;
 
 public abstract class LicensedRunnable implements Runnable {
