+++ tests/org.eclipse.passage.lic.base.tests/src/org/eclipse/passage/lic/internal/base/tests/requirements/RequirementsTest.java
@@ -20,11 +20,11 @@ import java.util.Collection;
 
 import org.eclipse.passage.lic.api.registry.StringServiceId;
 import org.eclipse.passage.lic.api.tests.fakes.requirements.FakeResolvedRequirements;
+import org.eclipse.passage.lic.base.access.Requirements;
 import org.eclipse.passage.lic.base.registry.ReadOnlyRegistry;
 import org.eclipse.passage.lic.internal.api.ServiceInvocationResult;
 import org.eclipse.passage.lic.internal.api.requirements.Requirement;
 import org.eclipse.passage.lic.internal.api.requirements.ResolvedRequirements;
-import org.eclipse.passage.lic.internal.base.access.Requirements;
 import org.eclipse.passage.lic.internal.base.diagnostic.code.NoRequirements;
 import org.eclipse.passage.lic.internal.base.diagnostic.code.NoServicesOfType;
 import org.junit.Test;
